Create packages/cache; lift shared MemoryCacheService (closes #290)
Creates @transformotion/cache with CacheService interface and MemoryCacheService
(in-memory + localStorage). Both apps now import the shared types and
MemoryCacheService from the package; SA's DynamoTTLCacheService stays app-local.
