+++ apps/budget-tracker/package.json
@@ -42,6 +42,7 @@
     "@radix-ui/react-toggle-group": "1.1.11",
     "@radix-ui/react-tooltip": "1.2.8",
     "@transformotion/api-client": "workspace:*",
+    "@transformotion/cache": "workspace:*",
     "@transformotion/ui-error-boundaries": "workspace:*",
     "@transformotion/ui-primitives": "workspace:*",
     "@transformotion/auth-client": "workspace:*",
