+++ apps/stock-analyser/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@transformotion/api-client": "workspace:*",
+    "@transformotion/cache": "workspace:*",
     "@transformotion/ui-error-boundaries": "workspace:*",
     "@transformotion/ui-primitives": "workspace:*",
     "@transformotion/auth-client": "workspace:*",
