+++ packages/cache/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@transformotion/cache",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./src/index.ts",
+  "scripts": {
+    "build": "tsc --noEmit",
+    "typecheck": "tsc --noEmit",
+    "lint": "eslint ."
+  },
+  "devDependencies": {
+    "@types/node": "^22",
+    "typescript": "*"
+  }
+}
